feat: extract and test openapi error conversion
